moved get_user_ids_from_usernames
